Made slight adjustments to PizzaBox
Adjusted some numbers to fit the new scale of things. Due to the boxes being half their size, all the numbers were off.

Also started figuring out a way to spawn broken boxes on death, but it probably needs the broken ones to have a script that organises their spawn.

Need to figure out a round about for when the pizza boxes are in a collider such as the spawn point. It interferes with the pizza box script.